feat(utils library): define @NotNullWhenSuchChildType() decorator
This decorator is wrapper for typeorm's @Check() decorator. Its mission is to restrict NOT NULL
constraint for child tables when they are defined with typeorm STI (@SingleTableInheritance() and
@ChildEntity() decorators). Because in reallity all childs tables are abstractions of one, it is
impossible define such constraint  only for part of rows. But it is possible compbine this
constraint with check for current row's "type" property. Of course it predict some conventions from
your side (see documentation about project in README files). But for example you should name your
properties in CamelCase, entity names should end with 'Entity' and for all childs should be
implicitly define 'type' name as enum, builded from their class's name without 'Entity' suffics. So
'OperatorEntity' should became 'Operator', etc.

BREAKING CHANGE: It is cool part of the code... but until we write tests for it, it always be little
risk. What if some bugs exists here. So if future will be some problems and any tests for  this part
of code.. please check it and write tests.